Tailable cursor demonstration
This commit demonstrates how easily Spring Data MongoDB allows to
expose a tailable cursor. All we need to do is to define an extra
operation on our repository and we can expose the `Flux` to get SSE
pretty easily.

We can use that on the client side to refresh the grap as new data comes
in.